Store hardcoded size grids as long for HIP

+++ core/src/impl/KokkosExp_IterateTileGPU.hpp
@@ -42,8 +42,8 @@ struct EmulateCUDADim3 {
 constexpr int mdrange_max_blocks_x = 2147483647;  // 2^31 - 1
 constexpr int mdrange_max_blocks   = 65535;       // 2^16 - 1
 #elif defined(KOKKOS_ENABLE_HIP)
-constexpr int mdrange_max_blocks_x = 4294967295;  // 2^32 - 1
-constexpr int mdrange_max_blocks   = 4294967295;  // 2^32 - 1
+constexpr long mdrange_max_blocks_x = 4294967295;  // 2^32 - 1
+constexpr long mdrange_max_blocks   = 4294967295;  // 2^32 - 1
 #elif defined(KOKKOS_ENABLE_SYCL)
 constexpr int mdrange_max_blocks_x = 4294967295;  // 2^32 - 1
 constexpr int mdrange_max_blocks   = 4294967295;  // 2^32 - 1
